fix: separate status enums
