+++ src/main/java/most/qms/models/GroupStatus.java
@@ -1,0 +1,8 @@
+package most.qms.models;
+
+public enum GroupStatus {
+    WAITING,
+    CALLED,
+    COMPLETE,
+    CANCELED,
+}
